Make NCBI taxa loader work for all <= species ranks
The loader was originally missing the forma and varietas ranks, and many
more have been added since it was written.

To future proof queries against more rank changes, the `species_or_below`
flag has been added to nodes to denode they are of species rank or lower, which
is the target of othe queries in question,
which are currently hard coded to various ranks.